Move some TenantInfo and NamespaceRowMapperInjector to core module.
