+++ core/src/main/java/com/alibaba/nacos/core/namespace/model/TenantInfo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ * Copyright 1999-2023 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.config.server.model;
+package com.alibaba.nacos.core.namespace.model;
 
 import java.io.Serializable;
 
 /**
  * TenantInfo.
+ *
+ * <p>Old name of {@link Namespace}</p>
  *
  * @author Nacos
  */
