+++ core/src/main/java/com/alibaba/nacos/core/persistence/DistributedDatabaseOperateImpl.java
@@ -137,7 +137,7 @@ import java.util.stream.Collectors;
  *               └───────────────────────────────────────────┘
  * </pre>
  *
- * TODO depend on Member and Cp protocol strongly, Waiting for addition split.
+ * <p>TODO depend on Member and Cp protocol strongly, Waiting for addition split.
  *
  * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
  */
